requirements.txt
PyYAML~=6.0.2
matplotlib~=3.8.2
discord.py~=2.4.0
python-dotenv~=1.1.0